Roles & permissions 1

+++ public_html/src/Controller/AdminController.php
@@ -49,44 +49,59 @@ class AdminController
     private $request;
 
 
+    /**
+     * User Controller
+     *
+     * @var UserController
+     */
+    private $userController;
+
+
     public function __construct()
     {
         $this->environment = new TwigRenderer();
         $this->renderer = $this->environment->getTwig();
         $this->postManager = new PostManager();
         $this->userManager = new UserManager();
+        $this->userController = new UserController();
         $this->session = new Session();
         $this->request =  \GuzzleHttp\Psr7\ServerRequest::fromGlobals();
     }
 
     public function index(): Response
     {
-        if (!empty($this->session->get('userID')) && !empty($this->session->get('username'))) 
-        {
-            $user = $this->userManager->read($this->session->get('userID'));
-            return new Response(200, [], $this->renderer->render('admin.html.twig', ['user' => $user]));
+        // only allow access to users who are both logged in and have admin role
+        if(!$this->userController->isLoggedIn()){
+            return new Response(301, ['Location' => 'login']);
         }
-        return new Response(200, [], $this->renderer->render('admin.html.twig'));
+        $user = $this->userManager->read($this->session->get('userID'));
+        return new Response(200, [], $this->renderer->render('admin.html.twig', ['user' => $user]));
     }
 
     // display the form to create a new Blog Post
     public function createPost(): Response
     {
+        // only allow access to users who are both logged in and have admin role
+        if(!$this->userController->isLoggedIn()){
+            return new Response(301, ['Location' => 'login']);
+        }
         return new Response(200, [], $this->renderer->render('create-post.html.twig'));
     }
     
     // send the request to Database
     public function addPost(): Response
     {
-        if (!empty($this->session->get('userID')) && !empty($this->session->get('username'))) 
-        {
-            $user = $this->userManager->read($this->session->get('userID'));
+        // only allow access to users who are both logged in and have admin role
+        if(!$this->userController->isLoggedIn()){
+            return new Response(301, ['Location' => 'login']);
         }
+        $user = $this->userManager->read($this->session->get('userID'));
         $title =  $this->request->getParsedBody()['title'];
         $content =  $this->request->getParsedBody()['content'];
         $slug =  $this->request->getParsedBody()['slug'];
+        $leadSentence = $this->request->getParsedBody()['leadSentence'];
         $author = $user->getId();
-        $newBlogPostId = $this->postManager->create($title, $content, $author, $slug);
+        $newBlogPostId = $this->postManager->create($title, $content, $author, $slug, $leadSentence);
         // TODO: next step 3 : sécurité (htmlspecialchars etc)
         $this->postManager->read($newBlogPostId);
         // redirect to Admin Blog Posts List
@@ -95,13 +110,16 @@ class AdminController
 
     public function showPosts(): Response
     {
+        // only allow access to users who are both logged in and have admin role
+        if(!$this->userController->isLoggedIn()){
+            return new Response(301, ['Location' => 'login']);
+        }
         $posts = $this->postManager->readAll();
         foreach ($posts as $post) {
             $authorID = $post->getCreated_By();
             $author = $this->userManager->read($authorID);
             $authors[$authorID] = $author;
         }
-
         return new Response(
             200,
             [],
@@ -118,16 +136,25 @@ class AdminController
     // form to edit a blog post
     public function editPost($id): Response
     {
+        // only allow access to users who are both logged in and have admin role
+        if(!$this->userController->isLoggedIn()){
+            return new Response(301, ['Location' => 'login']);
+        }
         $post = $this->postManager->read($id);
         return new Response(200, [], $this->renderer->render('edit-post.html.twig', ['post' => $post]));
     }
 
     public function updatePost($id): Response
     {
+        // only allow access to users who are both logged in and have admin role
+        if(!$this->userController->isLoggedIn()){
+            return new Response(301, ['Location' => 'login']);
+        }
         $title =  $this->request->getParsedBody()['title'];
         $content =  $this->request->getParsedBody()['content'];
         $slug =  $this->request->getParsedBody()['slug'];
-        $this->postManager->update($id, $title, $content, $slug);
+        $leadSentence = $this->request->getParsedBody()['leadSentence'];
+        $this->postManager->update($id, $title, $content, $slug, $leadSentence);
         // TODO: next step 3 : sécurIté (htmlspecialchars etc)
         // redirect to Admin Blog Posts List
         return new Response(301, ['Location' => '/admin/show-posts']);
@@ -135,6 +162,10 @@ class AdminController
 
     public function deletePost($id): Response
     {
+        // only allow access to users who are both logged in and have admin role
+        if(!$this->userController->isLoggedIn()){
+            return new Response(301, ['Location' => 'login']);
+        }
         $this->postManager->delete($id);
         // redirect to Admin Blog Posts List
         return new Response(302, ['Location' => '/admin/show-posts']);
@@ -144,21 +175,25 @@ class AdminController
 
     public function showUsers(): Response
     {
+        // only allow access to users who are both logged in and have admin role
+        if(!$this->userController->isLoggedIn()){
+            return new Response(301, ['Location' => 'login']);
+        }
         $users = $this->userManager->readAll();
-        $roles = $this->getRoles();
-        var_dump($users);
-        die();
+
         return new Response(200, [], $this->renderer->render('users.html.twig', ['users' => $users]));
     }
 
-    public function getRoles()
-    {
-        return $this->userManager->getAllRoles();
-    }
+    // public function getRoles()
+    // {
+    //     return $this->userManager->getAllRoles();
+    // }
 
-    public function assignedRoles($user, $roles)
-    {
-        // if(str_contains($user->getRoles(), $r))
-        //TODO:
-    }
+    // public function assignedRoles($user, $roles)
+    // {
+    //     // if(str_contains($user->getRoles(), $r))
+    //     //TODO:
+    // }
+
+
 }
